Update pipeline
Split update of upgrade-opsman-aws into its own file.

The config for this requires values that are the same key but with a
different value than other pipelines, like the AWS_ACCESS_KEY_ID.

Instead of trying to make the generic update_pipeline do flips to make
it work, split into a new file for simplicity.