Split put of pcf-pipelines-release into src/tags
In the event that new code is pushed while the publish-rc job runs, the
publish-rc job will fail to push the updated source code.

To avoid this problem we push the source before pushing the tag, since
it's safe to push the tag after the source succeeds and we don't want to
push the tag if the source push fails.

Configuring the resource to do a rebase is not safe in this case because
the resource version must be stamped in all of the pipeline.yml files,
and that process would have to be repeated to ensure any new pipelines
have the correct resource version stamped in them.