Modify Service Network config
The service network should be used for none dynamic services. If this value is set to true you can not put regular service tiles on the network as should be allowed I believe. The Dynamic service network is the one that should be checked true as a service network bosh does not control.