Make the GitHub release usable outside of Pivnet
This change replaces all instances of getting pcf-pipelines-tarball (the
Pivnet resource) with a get of pcf-pipelines (the GitHub resource).

The CI pipeline has been changed to no longer stamp the Pivnet release
version accordingly; it now uses
[yaml-patch](github.com/krishicks/yaml-patch) to create the release
version of pcf-pipelines that is distributed on Pivnet.

We also no longer modify the repository as part of shipping the Pivnet
release, though we do still tag the commit that was shipped.

The downside to using `yaml-patch` to modify the pipelines is that, due
to the random sort of Golang maps, the release pipelines are no longer
as pretty as they once were. The `jobs:` key, for example, may end up in
a different spot than where it is in the version of pcf-pipelines on
GitHub, or the `name:` key of jobs may end up at the bottom of the job
declaration rather than at the top. None of this impacts how the
pipeline works; just how easy it is to read as an end-user of the Pivnet
release.

The major upside of using `yaml-patch` is that as an end-user you are
free to create ops files similar to the
`operations/use-pivnet-release.yml` ops file in
`pcf-pipelines/operations` to add custom extensions to the pipelines.
This facility can be used for converting the pipelines to be offline
only, adding or removing tasks to jobs, adding jobs, etc.

It also, crucially, means you're now able to have your alterations
separate from pcf-pipelines itself, which will facilitate being able to
upgrade pcf-pipelines.

Note that due to the nature of the `{{placeholder}}` values being
invalid YAML, prior to running the pipeline through `yaml-patch` you
need to use something like `sed` to replace instances of
`{{placeholders}}` with quoted versions of same, such as
`"{{placeholders}}"`, to make them valid YAML, and then unwrap the
placeholders after applying `yaml-patch`. An example of this workflow
exists in the aforementioned `operations/use-pivnet-release.yml` ops
file.

In the future this modifying of placeholders won't be necessary as 1)
Concourse will adopt double parens instead of double curly braces and/or
2) the ops file functionality will be rolled into the `fly` CLI, making
using `yaml-patch` external to `fly` unnecessary.

[#144034687]