Pass OPSMAN_URI to upgrade-ops-manager pipeline
Pass OPSMAN_URI to `upgrade-ops-manager pipeline` so `deploy-opsman-vm` task uses it instead of OPSMAN_IP when checking if new OpsMan web server is up and running. This makes the pipeline work for all customers, including the ones where the OpsMan domain is NAT'ed and its VM IP address is not reachable from Concourse.