Fix upload-product-and-stemcell when no stemcell
Due to the `-e` nature of this script, if the bash call failed it would
exit early. We expect that it might fail, so instead of doing that, and
also instead of just wrapping the call in `set +e` and `set -e`, use
`jq` instead of grep as it has the most expressiveness.