Use jq in config-opsdir/task.sh
* config-opsdir/task.sh now correctly handles trusted certificates when
  provided as multiline YAML strings

Params changes:

* Remove generate_vm_passwords for install-pcf/vsphere.

Notes:

Due to an issue with how spaces are encoded in Golang when in a query
param (as `+`), `om-linux configure-bosh --az-configuration` fails when
the availability zones have spaces in them.

To mitigate this we continue to use `om-linux curl`, which seems to pass
along the JSON without modification.