Use semver-ish sort in upload-product-and-stemcell
Previously this relied on `jq`'s internal sort method, which when
applied to strings does an ASCII-based sort, meaning `"3312.9" >
"3312.33"`.

This is no good when trying to get the latest stemcell.

This change makes it so we take the array of stemcell versions and then
split them up by `.` into multiple arrays of numbers. We can then
compare the arrays of numbers to find the greatest.

[#149930884]