Add apply-updates pipeline
This can be used to replace upgrade-ert and be extended to replace other
upgrade-tile pipelines given the an opfile like the example,
add-more-tile-uploads-to-apply-updates.yml.example.

This also adds every new stemcell for the configured iaas to PCF
Operations Manager.

The `apply-updates` job stages the latest version of all the uploaded
tiles if and only if there is a version of the tile which has not yet
been staged. An opfile can be used to add a regulator job to that
automatically triggers `apply-updates`.

[#150563661]