Parameterize globs entry for upgrade-tile pipeline
A new parameter has been added to the upgrade-tile pipeline
to allow for an easy change of the default glob "*pivotal"
for a specific tile and to allow that pipeline to also be
used for upgrade of the ERT tile while avoiding the download
of the SRT file (by using cf*pivotal globs regexp for it).
The default glob for the currently generated upgrade-ert
pipeline has also bee updated accordingly.

[#152317070]