add comment listing ami NAT ids > > The AMIs for NAT instances by the region they are in are important for the terraforming process. This list is maintained by the git ci -m 'add comment listing ami NAT ids
The AMIs for NAT instances by the region they are in are important for the
terraforming process. This list is maintained by the infrastructure team, which
they periodically update when needed. We are using their work for reference.

We could not include the original PR as there were merge conflicts. We'd like to thank @cpage-pivotal for the help

[#154174869]

Signed-off-by: David Timm <dtimm@pivotal.io>