replace regex version with CHANGEME
We have people reporting that the regex is not respecting a certain version of Ops Manager and PCF.
It has been changed to be a CHANGEME with a comment that explicit says the regex for the correct verison.
It is better to make the user think about it, so they know what they are getting into.

[#156559165]

Signed-off-by: Huan Wang <huwang@pivotal.io>