Add connect and request timeout configurations.
In our environment, we have been experiencing several connection related
timeouts while using the om cli utility during the apply-changes task.
As of om 0.37.0, it provides the ability to configure both a connect and
a request timeout. This change allows both of those to be user
configurable within external pipelines.